Finished spliting RegistroClases.
