+++ FrontEnd/src/Components/Registro/RegistroClasesBody.jsx
@@ -1,37 +1,13 @@
-import Box from '@mui/material/Box';
-import React, { useState } from 'react';
-import {
-  Button,
-  Card,
-  CardContent,
-  Typography,
-  TextField,
-  MenuItem,
-} from '@mui/material';
-import Autocomplete from '@mui/material/Autocomplete';
-import { DataGrid } from '@mui/x-data-grid';
-import Modal from '@mui/material/Modal';
-import SearchIcon from '@mui/icons-material/Search';
-import ClaseModal from '../Clase/ClaseModal';
-import MiRegistro from './MiRegistro';
-import Clase from '../Clase/Clase';
+import React from 'react';
 import {
   getNivel, getHorario, getProfesor, getCupo,
 } from '../../utils/utilFunctions';
 import ButtonActionsInscripcion from './ButtonActionsInscripcion';
+import RegistroClasesTable from './RegistroClasesTable';
 
 function RegistroClasesBody({
   handleClick, filteredClasses, classNames,
 }) {
-  const [items, setItems] = useState([]);
-  const [currentClase, setCurrentClase] = useState();
-  const [openMoreInfo, setOpenMoreInfo] = useState(false);
-
-  const handleMoreInfo = (clase) => {
-    setCurrentClase(clase);
-    setOpenMoreInfo(!openMoreInfo);
-  };
-
   const columns = [
     {
       field: 'clavePeriodo',
@@ -100,193 +76,12 @@ function RegistroClasesBody({
   ];
 
   return (
-    <>
-      <Box
-        sx={{
-          textAlign: 'center',
-          width: '100%',
-          paddingX: '20px',
-          paddingBottom: '10px',
-          overflowY: 'scroll',
-          display: { xs: 'block', md: 'none' },
-        }}
-      >
-        {filteredClasses.length !== 0 ? (
-          filteredClasses.map((e) => (
-            <Clase
-              handleClick={handleClick}
-              handleMoreInfo={handleMoreInfo}
-              key={e._id}
-              clase={e}
-            />
-          ))
-        ) : (
-          <Box
-            sx={{
-              height: '100vh',
-              display: 'flex',
-              alignContent: 'center',
-              justifyContent: 'center',
-              flexWrap: 'wrap',
-            }}
-          >
-            <Typography variant="h3" component="div" textAlign="center">
-              No hay clases disponibles por el momento.
-            </Typography>
-          </Box>
-        )}
-      </Box>
-      <Box
-        sx={{
-          display: { xs: 'none', md: 'flex' },
-          justifyContent: 'center',
-          alignItems: 'center',
-          flexDirection: 'column',
-        }}
-      >
-        <Box sx={{ display: 'flex' }}>
-          <Card
-            sx={{
-              textAlign: 'center',
-              ml: 1,
-              my: 2,
-              display: 'flex',
-            }}
-          >
-            <SearchIcon
-              color="primary"
-              width="2em"
-              height="2em"
-              sx={{ alignSelf: 'center', ml: 0.5 }}
-            />
-            <CardContent
-              sx={{
-                display: 'flex',
-                flexDirection: 'row',
-                flexWrap: 'wrap',
-                alignItems: 'center',
-                '& .MuiTextField-root': { m: 1, width: '25ch' },
-                p: 1,
-              }}
-            >
-              <Autocomplete
-                disablePortal
-                options={classNames}
-                onChange={(_e, newvalue) => {
-                  setItems([
-                    {
-                      columnField: 'nombre_curso',
-                      operatorValue: 'contains',
-                      value: newvalue,
-                    },
-                  ]);
-                }}
-                onInputChange={(_e, newvalue) => {
-                  setItems([
-                    {
-                      columnField: 'nombre_curso',
-                      operatorValue: 'contains',
-                      value: newvalue,
-                    },
-                  ]);
-                }}
-                renderInput={(params) => (
-                  <TextField {...params} label="Curso" />
-                )}
-              />
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Nivel"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'nivel',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-                select
-              >
-                {[
-                  '',
-                  'Desde cero',
-                  'Con bases',
-                  'Intermedio',
-                  'Avanzado',
-                ].map((e) => (
-                  <MenuItem value={e} key={e}>
-                    {e}
-                  </MenuItem>
-                ))}
-              </TextField>
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Periodo"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'clavePeriodo',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-              />
-              <TextField
-                style={{ fontFamily: 'arial' }}
-                label="Modalidad"
-                onChange={(e) => {
-                  setItems([
-                    {
-                      columnField: 'modalidad',
-                      operatorValue: 'contains',
-                      value: e.target.value,
-                    },
-                  ]);
-                }}
-              />
-            </CardContent>
-          </Card>
-          <MiRegistro />
-        </Box>
-        <Box
-          sx={{
-            m: 2,
-            display: 'flex',
-            width: '90%',
-            height: 600,
-            minWidth: '548px',
-            '& .theme--ListaEspera': {
-              bgcolor: 'lightyellow',
-            },
-            '& .theme--Inscrito': {
-              bgcolor: 'lightgreen',
-            },
-          }}
-        >
-          <DataGrid
-            sx={{ flexGrow: 1 }}
-            rows={filteredClasses}
-            columns={columns}
-            disableSelectionOnClick
-            getRowId={(row) => row._id}
-            getRowHeight={() => 'auto'}
-            filterModel={{
-              items,
-            }}
-            getRowClassName={(params) => `theme--${params.row.status}`}
-          />
-        </Box>
-        <Modal
-          open={openMoreInfo}
-          onClose={() => setOpenMoreInfo(!openMoreInfo)}
-          sx={{ overflowY: 'scroll' }}
-        >
-          <ClaseModal clase={currentClase} />
-        </Modal>
-      </Box>
-    </>
+    <RegistroClasesTable 
+      handleClick={handleClick}
+      filteredClasses={filteredClasses}
+      classNames={classNames}
+      columns={columns}
+    />
   );
 }
